Cpp cleanup
